add pledges and basic done reporting
(about to get on plane so this is still a little rough)

the idea is that a given person can commit (i.e. pledge) to do a given goal
once per 31 day time period, and then say whether they did it. this way if they
are on more than one team, they can get credit for doing the same goal for each
team, but we'll still always account for things in terms of what a given person
did/didn't do.