better release flow
we should just use `image.release` now. this will push a docker image
tagged with the hash of the repo state and then configure the deployment
to use that image version. this allows convenient rollbacks and lets you
track changes through the UI.

if you pass your own version that looks semver-like, this will still get
set on the git repo and pushed upstream